+++ sdk/storage/azure-storage-files-shares/src/share_file_client.cpp
@@ -407,7 +407,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
         m_shareFileUrl, *m_pipeline, context, protocolLayerOptions);
   }
 
-  Azure::Response<Models::GetShareFilePropertiesResult> ShareFileClient::GetProperties(
+  Azure::Response<Models::ShareFileProperties> ShareFileClient::GetProperties(
       const GetShareFilePropertiesOptions& options,
       const Azure::Core::Context& context) const
   {