+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -11,6 +11,12 @@
 
 - Changed the return type of `StartCopy` API from a `Response<T>` to the particular `Operation<T>` type called `StartCopyShareFileOperation` directly.
 - String conversion functions of extensible enums were renamed from `Get()` to `ToString()`.
+- Renamed `GetSharePropertiesResult` to `ShareProperties`.
+- Renamed `GetShareStatisticsResult` to `ShareStatistics`.
+- Renamed `GetShareDirectoryPropertiesResult` to `ShareDirectoryProperties`.
+- Renamed `GetShareFilePropertiesResult` to `ShareFileServiceProperties`.
+- `ShareClient::GetPermission` now returns `Azure::Response<std::string>`.
+- `ShareServiceClient::GetProperties` now returns `Azure::Response<Models::ShareFileServiceProperties>`.
 
 ## 12.0.0-beta.8 (2021-02-12)